Add(minjae/refactor): 마켓 리팩토링 #81
